Moved lifetoken test to gametest
