+++ src/test/java/inf112/skeleton/app/PlayerTest.java
@@ -1,28 +1,19 @@
 package inf112.skeleton.app;
 
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.backends.headless.HeadlessApplication;
-import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.math.Vector2;
 import inf112.skeleton.app.enums.Direction;
 import org.junit.Before;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
-import static org.mockito.Mockito.mock;
+import static org.junit.Assert.assertTrue;
 
 public class PlayerTest {
 
     private Player player;
-    private RallyGame game;
 
     @Before
     public void setUp() {
-        Gdx.gl = mock(GL20.class);
-        //Make a headless application in order to initialize the board. Does not show.
-        new HeadlessApplication(new EmptyApplication());
-        this.game = new RallyGame();
-        this.game.setupGame("assets/maps/Risky_Exchange.tmx");
         Vector2 pos = new Vector2(0,0);
         player = new Player(pos, 1);
     }
@@ -78,15 +69,11 @@ public class PlayerTest {
     }
 
     @Test
-    public void lifeDecreasedWhenCollectedTenDamageTokensTest() {
-        Board board = game.getBoard();
-        board.addPlayer(player);
-        int livesBefore = player.getLifeTokens();
-        for (int takeDamage = 1; takeDamage <= 10; takeDamage++) {
-            player.handleDamage();
+    public void noLivesLeftPlayerIsDeadTest() {
+        for (int livesTaken = 1; livesTaken <=3; livesTaken++) {
+            player.decrementLifeTokens();
         }
-        game.decreaseLives();
-        assertEquals(livesBefore-1, player.getLifeTokens());
+        assertTrue(player.isDead());
     }
 
 }
